requirements.txt
requests
tweepy
pillow
google-cloud-storage